% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_pdf_header.R
\name{create_pdf_header}
\alias{create_pdf_header}
\title{Create pdf header.tex file}
\usage{
create_pdf_header(
  author = "Author",
  title = "Title",
  description = "Description",
  email = "email@email.com",
  slogan = "R report",
  created_on = "Created on",
  bg = "Background_dark_topdown_ThinkR.png",
  bg.title = "Background_Title_dark_ThinkR.png",
  link.col = "blue",
  section.color = "forestgreen",
  main.col = "#192ac7",
  out.header = file.path(tempdir(), "header.tex"),
  company = "Company",
  company_url = "https://company.fr"
)
}
\arguments{
\item{author}{Name of the author of the report}

\item{title}{Title of the report}

\item{description}{Short description of the report}

\item{email}{Email for contact}

\item{slogan}{Sentence that}

\item{created_on}{Allow translation for "Created on"}

\item{bg}{path to PNG image used a main background}

\item{bg.title}{path to PNG image used a Report Title background}

\item{link.col}{Color in the R language (name or hex)}

\item{section.color}{Color of section titles}

\item{main.col}{Color of the title on the title page and Company name in header}

\item{out.header}{Path where to save header.tex}

\item{company}{company name}

\item{company_url}{company_url}
}
\description{
Create pdf header.tex file
}
