% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_pdf_header.R
\name{pdf_header}
\alias{pdf_header}
\title{Create content of latex header.tex file}
\usage{
pdf_header(
  author,
  title,
  email,
  description,
  slogan = "R report",
  created_on = "Created on",
  bg = "Background_dark_topdown_ThinkR.png",
  bg.title = "Background_Title_dark_ThinkR.png",
  link.col = "#FF8000",
  section.color = "#2447c4",
  main.col = "#192ac7",
  company,
  company_url
)
}
\arguments{
\item{author}{Name of the author of the report}

\item{title}{Title of the report}

\item{email}{Email for contact}

\item{description}{Short description of the report}

\item{slogan}{Sentence that}

\item{created_on}{Allow translation for "Created on"}

\item{bg}{path to PNG image used a main background}

\item{bg.title}{path to PNG image used a Report Title background}

\item{link.col}{Color in the R language (name or hex)}

\item{section.color}{Color of section titles}

\item{main.col}{Color of the title on the title page and Company name in header}

\item{company}{company name}

\item{company_url}{company_url}
}
\description{
Create content of latex header.tex file
}
