% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_pdf_header.R
\name{pdf_headerfooter}
\alias{pdf_headerfooter}
\title{Create header and footer of the pages}
\usage{
pdf_headerfooter(author, email, slogan = "R report",
  created_on = "Created on", company = "ThinkR",
  company_url = "https://thinkr.fr")
}
\arguments{
\item{author}{Name of the author of the report}

\item{email}{Email for contact}

\item{slogan}{Sentence that}

\item{created_on}{Allow translation for "Created on"}

\item{company}{company name}

\item{company_url}{company_url}
}
\description{
Create header and footer of the pages
}
