% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_pdf_header.R
\name{pdf_layout}
\alias{pdf_layout}
\title{Pdf Layout}
\usage{
pdf_layout(
  bg = "Background_dark_topdown_ThinkR.png",
  bg.title = "Background_Title_dark_ThinkR.png",
  link.col = "#FF8000",
  main.col = "#192ac7"
)
}
\arguments{
\item{bg}{path to PNG image used a main background}

\item{bg.title}{path to PNG image used a Report Title background}

\item{link.col}{Color in the R language (name or hex)}

\item{main.col}{Color of the title on the title page and Company name in header}
}
\description{
Pdf Layout
}
