% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_pdf_header.R
\name{pdf_sections}
\alias{pdf_sections}
\title{Format titles}
\usage{
pdf_sections(section.color = "#2447c4")
}
\arguments{
\item{section.color}{Color of section titles}
}
\description{
Format titles
}
