% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_pdf_header.R
\name{pdf_title}
\alias{pdf_title}
\title{Title page of the report}
\usage{
pdf_title(author, title, description, company, company_url)
}
\arguments{
\item{author}{Name of the author of the report}

\item{title}{Title of the report}

\item{description}{Short description of the report}

\item{company}{company name}

\item{company_url}{company_url}
}
\description{
Title page of the report
}
