% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_pdf_header.R
\name{prepare_yaml_for_knit}
\alias{prepare_yaml_for_knit}
\title{Prepare yaml for knitting}
\usage{
prepare_yaml_for_knit(rmd.path, fig_caption = TRUE, keep_tex = FALSE,
  number_sections = TRUE, toc = TRUE, lang = "en",
  out_format = c("pdf_document2", "pdf_book"))
}
\arguments{
\item{rmd.path}{Path of the Rmd file to be knit}

\item{fig_caption}{\code{TRUE} to render figures with captions}

\item{keep_tex}{Keep the intermediate tex file used in the conversion to PDF}

\item{number_sections}{\code{TRUE} to number section headings}

\item{toc}{\code{TRUE} to include a table of contents in the output}

\item{lang}{Document language code (e.g. "en", "es", "fr", "pt-BR")}

\item{out_format}{output pdf format among "pdf_book", "pdf_document2"}
}
\description{
Prepare yaml for knitting
}
\details{
Document dimensions and margins are fixed as linked with background image
}
